libm : cos : rewrite cos and sin

+++ libm/generic/cosf.c
@@ -1,21 +1,21 @@
 #include <math.h>
 
 // constants taken from FreeBSD
-static const double c[] = {
-	 4.16666666666666019037e-02,
-	-1.38888888888741095749e-03,
-	 2.48015872894767294178e-05,
-	-2.75573143513906633035e-07,
-	 2.08757232129817482790e-09,
-	-1.13596475577381996161e-11,
+static const float c[] = {
+	 4.16666666666666019037e-02f,
+	-1.38888888888741095749e-03f,
+	 2.48015872894767294178e-05f,
+	-2.75573143513906633035e-07f,
+	 2.08757232129817482790e-09f,
+	-1.13596475577381996161e-11f,
 };
 
-static double cos_core(double x) {
-	double x2 = x * x;
-	double x4 = x2 * x2;
+static float cos_core(float x) {
+	float x2 = x * x;
+	float x4 = x2 * x2;
 
 	// minimax polyminal using Horner's Method
-	double res = c[5];
+	float res = c[5];
 	for (int i=4; i>=0; i--) {
 		res = res * x2 + c[i];
 	}
@@ -24,32 +24,32 @@ static double cos_core(double x) {
 }
 
 // constants taken from FreeBSD
-static const double s[] = {
-	-1.66666666666666324348e-01,
-	 8.33333333332248946124e-03,
-	-1.98412698298579493134e-04,
-	 2.75573137070700676789e-06,
-	-2.50507602534068634195e-08,
-	 1.58969099521155010221e-10,
+static const float s[] = {
+	-1.66666666666666324348e-01f,
+	 8.33333333332248946124e-03f,
+	-1.98412698298579493134e-04f,
+	 2.75573137070700676789e-06f,
+	-2.50507602534068634195e-08f,
+	 1.58969099521155010221e-10f,
 };
 
-static double sin_core(double x) {
-	double x2 = x * x;
-	double x3 = x2 * x;
+static float sin_core(float x) {
+	float x2 = x * x;
+	float x3 = x2 * x;
 
 	// minimax polyminal using Horner's Method
-	double res = s[5];
+	float res = s[5];
 	for (int i=4; i>=0; i--) {
 		res = res * x2 + s[i];
 	}
 	return x + x3 * res;
 }
 
-double cos(double x) {
+float cosf(float x) {
 	if (isnan(x)) return x;
 	if (isinf(x)) return NAN;
-	x = fmod(x, 2.0 * M_PI);
-	x = fabs(x);
+	x = fmodf(x, 2.0 * M_PI);
+	x = fabsf(x);
 
 	if (x <= M_PI_4) {
 		return cos_core(x);
@@ -64,17 +64,17 @@ double cos(double x) {
 	}
 }
 
-double sin(double x) {
+float sinf(float x) {
 	if (isnan(x)) return x;
 	if (isinf(x)) return NAN;
 
-	double sign = 1.0;
+	float sign = 1.0;
 	if (x < 0) {
 		x = -x;
 		sign = -1.0;
 	}
 
-	x = fmod(x, 2.0 * M_PI);
+	x = fmodf(x, 2.0 * M_PI);
 
 	if (x <= M_PI_4) {
 		return sin_core(x) * sign;
